feat: Response 값 Int로 변경

+++ Backend/kidswallet/src/main/java/com/e201/kidswallet/togetherrun/dto/RegularDepositResponseDto.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @Builder
 @AllArgsConstructor
 public class RegularDepositResponseDto {
-    private BigDecimal amount;
+    private int amount;
     private short depositDay;
     private LocalDate startDate;
     private LocalDate endDate;
